Added EmailUtil

+++ fitnesstrackertest-Gradle/src/main/java/com/nischitha/spring/fitnesstrackertest/repos/WorkoutReopository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.nischitha.spring.fitnesstrackertest.entities.Workout;
 
-public interface WorkoutReopository extends JpaRepository<Workout, Long> {
+public interface WorkoutReopository extends JpaRepository<Workout, Integer> {
 
 }
